move as_cwd into model_setup; remove scripting module
