+++ tests/modelmetadata_test.py
@@ -30,7 +30,7 @@ class BarBaz:
 def test_search_paths(model):
     assert ModelMetadata.search_paths(model) == (
         "FooBar",
-        os.path.join(sys.prefix, "share", "csdms", "FooBar")
+        os.path.join(sys.prefix, "share", "csdms", "FooBar"),
     )
 
 
@@ -39,7 +39,7 @@ def test_search_paths_with_metadata(model):
     assert ModelMetadata.search_paths(model) == (
         os.path.join(os.path.dirname(__file__), "FooBar"),
         "BarBaz",
-        os.path.join(sys.prefix, "share", "csdms", "BarBaz")
+        os.path.join(sys.prefix, "share", "csdms", "BarBaz"),
     )
